// src/app/modules/Faculty/faculty.interface.ts
/* eslint-disable no-unused-vars */
import { Model, Types } from "mongoose";

export type TGender = 'male' | 'female' | 'other';
export type TBloodGroup =
    | 'A+'
    | 'A-'
    | 'B+'
    | 'B-'
    | 'AB+'
    | 'AB-'
    | 'O+'
    | 'O-';

export type TUserName = {
    firstName: string;
    middleName: string;
    lastName: string;
};

export type TFaculty = {
    id: string;
    user: Types.ObjectId;
    designation: string;
    name: TUserName;
    gender: TGender;
    dateOfBirth?: Date;
    email: string;
    contactNo: string;
    emergencyContactNo: string;
    bloodGroup?: TBloodGroup;
    presentAddress: string
    permanentAddress: string
    profileImage?: string;
    academicDepartment: Types.ObjectId;
    academicFaculty: Types.ObjectId;
    isDeleted: boolean,
};

export interface FacultyModel extends Model<TFaculty>{
    isUserExists(id: string): Promise<TFaculty | null>
};