// src/app/modules/Admin/admin.interface.ts
/* eslint-disable no-unused-vars */
import { Model, Types } from 'mongoose';

export type TGender = 'male' | 'female' | 'other';
export type TBloodGroup =
    | 'A+'
    | 'A-'
    | 'B+'
    | 'B-'
    | 'AB+'
    | 'AB-'
    | 'O+'
    | 'O-';

export type TUserName = {
    firstName: string;
    middleName: string;
    lastName: string;
};

export type TAdmin = {
    id: string;
    user: Types.ObjectId;
    designation: string;
    name: TUserName;
    gender: TGender;
    dateOfBirth?: Date;
    email: string;
    contactNo: string;
    emergencyContactNo: string;
    bloodGroup?: TBloodGroup;
    presentAddress: string;
    permanentAddress: string;
    profileImage?: string;
    isDeleted: boolean;
};

export interface AdminModel extends Model<TAdmin> {
    isUserExists(id: string): Promise<TAdmin | null>;
}